python/olaf_discord_files/pyproject.toml
[tool.poetry]
name = "olaf-discord-files"
version = "0.1.0"
description = ""
authors = ["StarrFox <starrfox6312@gmail.com>"]
readme = "README.md"
packages = [{include = "olaf_discord_files"}]

[tool.poetry.scripts]
bot = "olaf_discord_files.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
